pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "applycopilot"
version = "0.1.0"
description = "Local-first AI-powered job & research internship tracker"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
authors = [{ name = "Harsh", email = "harsh06264@gmail.com" }]

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",

    # Config
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Auth
    "argon2-cffi>=23.1.0",
    "python-jose[cryptography]>=3.3.0",

    # Email
    "imapclient>=3.0.1",
    "mail-parser>=3.15.0",

    # Resume
    "python-docx>=1.1.0",

    # AI / Ollama
    "ollama>=0.2.0",
    "httpx>=0.27.0",

    # Utilities
    "python-dateutil>=2.9.0",
    "gitpython>=3.1.43",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "httpx>=0.27.0",
]

[project.scripts]
applycopilot = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "ai*", "db*", "api*", "cli*", "config*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true
